+++ mjpeg_streamer/__init__.py
@@ -1,4 +1,5 @@
-from .mjpeg_streamer import CustomStream, ManagedStream, MjpegServer, Stream
+from .server import MjpegServer
+from .stream import CustomStream, ManagedStream, Stream
 
 __all__ = ["CustomStream", "ManagedStream", "MjpegServer", "Stream"]
 __version__ = "2024.2.8"
